+++ Api.h
@@ -6,6 +6,6 @@ class Api {
   public:
     static int dispatch(State *state, uint8_t *payload, int length);
   private:
-    static int duration(uint8_t *recvBuffer, uint8_t length, uint8_t offset);
+    static uint16_t duration(uint8_t *recvBuffer, uint8_t length, uint8_t offset);
 };
 #endif
